let the parser pull tokens on demand
This allows the tokenizer to be lazy and only to produce a next token when the parser needs it. This reduces the amount of memory used for buffing all tokens before starting to convert them into segments.

This is a BC break, because the API of the TokenizerInterface changes: getTokens no longer returns an array, but any iterable.